fix(client-ec2): update ec2 model
