+++ clients/client-ec2/commands/CancelReservedInstancesListingCommand.ts
@@ -24,7 +24,7 @@ export type CancelReservedInstancesListingCommandOutput = CancelReservedInstance
  * <p>Cancels the specified Reserved Instance listing in the Reserved Instance Marketplace.</p>
  *          <p>For more information, see
  *         <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-market-general.html">Reserved Instance Marketplace</a>
- *         in the <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ *         in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class CancelReservedInstancesListingCommand extends $Command<
   CancelReservedInstancesListingCommandInput,
