+++ clients/client-ec2/commands/CreateFleetCommand.ts
@@ -21,7 +21,7 @@ export type CreateFleetCommandOutput = CreateFleetResult & __MetadataBearer;
  * <p>Launches an EC2 Fleet.</p>
  *          <p>You can create a single EC2 Fleet that includes multiple launch specifications that vary by
  *          instance type, AMI, Availability Zone, or subnet.</p>
- *          <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-fleet.html">Launching an EC2 Fleet</a> in the <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ *          <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-fleet.html">Launching an EC2 Fleet</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class CreateFleetCommand extends $Command<
   CreateFleetCommandInput,
