+++ clients/client-ec2/commands/CreatePlacementGroupCommand.ts
@@ -30,7 +30,7 @@ export type CreatePlacementGroupCommandOutput = CreatePlacementGroupResult & __M
  *             partitions, where instances in one partition do not share the same hardware with
  *             instances in another partition.</p>
  *         <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/placement-groups.html">Placement groups</a> in the
- *                 <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ *                 <i>Amazon EC2 User Guide</i>.</p>
  */
 export class CreatePlacementGroupCommand extends $Command<
   CreatePlacementGroupCommandInput,
