+++ clients/client-ec2/commands/DeleteFleetsCommand.ts
@@ -44,7 +44,7 @@ export type DeleteFleetsCommandOutput = DeleteFleetsResult & __MetadataBearer;
  *          </ul>
  *
  *          <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/manage-ec2-fleet.html#delete-fleet">Deleting an EC2
- *          Fleet</a> in the <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ *          Fleet</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class DeleteFleetsCommand extends $Command<
   DeleteFleetsCommandInput,
