+++ clients/client-ec2/commands/DescribeFleetInstancesCommand.ts
@@ -22,6 +22,7 @@ export type DescribeFleetInstancesCommandOutput = DescribeFleetInstancesResult &
 
 /**
  * <p>Describes the running instances for the specified EC2 Fleet.</p>
+ *          <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-fleet.html#monitor-ec2-fleet">Monitoring your EC2 Fleet</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class DescribeFleetInstancesCommand extends $Command<
   DescribeFleetInstancesCommandInput,
