+++ clients/client-ec2/commands/DescribeFleetsCommand.ts
@@ -19,6 +19,7 @@ export type DescribeFleetsCommandOutput = DescribeFleetsResult & __MetadataBeare
 
 /**
  * <p>Describes the specified EC2 Fleets or all of your EC2 Fleets.</p>
+ *          <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-fleet.html#monitor-ec2-fleet">Monitoring your EC2 Fleet</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class DescribeFleetsCommand extends $Command<
   DescribeFleetsCommandInput,
