+++ clients/client-ec2/commands/DescribeHostReservationOfferingsCommand.ts
@@ -26,8 +26,8 @@ export type DescribeHostReservationOfferingsCommandOutput = DescribeHostReservat
  * 			offerings that might not match the instance family and Region of your Dedicated Hosts.
  * 			When purchasing an offering, ensure that the instance family and Region of the offering
  * 			matches that of the Dedicated Hosts with which it is to be associated. For more
- * 			information about supported instance types, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/dedicated-hosts-overview.html">Dedicated Hosts
- * 				Overview</a> in the <i>Amazon Elastic Compute Cloud User Guide</i>. </p>
+ * 			information about supported instance types, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/dedicated-hosts-overview.html">Dedicated Hosts</a>
+ *             in the <i>Amazon EC2 User Guide</i>. </p>
  */
 export class DescribeHostReservationOfferingsCommand extends $Command<
   DescribeHostReservationOfferingsCommandInput,
