+++ clients/client-ec2/commands/DescribeInstanceCreditSpecificationsCommand.ts
@@ -44,8 +44,7 @@ export type DescribeInstanceCreditSpecificationsCommandOutput = DescribeInstance
  *             IDs in the affected zone, or do not specify any instance IDs at all, the call fails. If
  *             you specify only instance IDs in an unaffected zone, the call works normally.</p>
  *         <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/burstable-performance-instances.html">Burstable
- *                 performance instances</a> in the <i>Amazon Elastic Compute Cloud User
- *                 Guide</i>.</p>
+ *                 performance instances</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class DescribeInstanceCreditSpecificationsCommand extends $Command<
   DescribeInstanceCreditSpecificationsCommandInput,
