+++ clients/client-ec2/commands/DescribeInstanceStatusCommand.ts
@@ -31,24 +31,21 @@ export type DescribeInstanceStatusCommandOutput = DescribeInstanceStatusResult &
  *                     <b>Status checks</b> - Amazon EC2 performs status
  *                     checks on running EC2 instances to identify hardware and software issues. For
  *                     more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/monitoring-system-instance-status-check.html">Status checks for your instances</a> and <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/TroubleshootingInstances.html">Troubleshooting
- *                         instances with failed status checks</a> in the <i>Amazon Elastic
- *                         Compute Cloud User Guide</i>.</p>
+ *                         instances with failed status checks</a> in the <i>Amazon EC2 User Guide</i>.</p>
  *             </li>
  *             <li>
  *                 <p>
  *                     <b>Scheduled events</b> - Amazon EC2 can schedule
  *                     events (such as reboot, stop, or terminate) for your instances related to
  *                     hardware issues, software updates, or system maintenance. For more information,
- *                     see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/monitoring-instances-status-check_sched.html">Scheduled events for your instances</a> in the <i>Amazon Elastic
- *                         Compute Cloud User Guide</i>.</p>
+ *                     see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/monitoring-instances-status-check_sched.html">Scheduled events for your instances</a> in the <i>Amazon EC2 User Guide</i>.</p>
  *             </li>
  *             <li>
  *                 <p>
  *                     <b>Instance state</b> - You can manage your instances
  *                     from the moment you launch them through their termination. For more information,
  *                     see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-lifecycle.html">Instance
- *                         lifecycle</a> in the <i>Amazon Elastic Compute Cloud User
- *                         Guide</i>.</p>
+ *                         lifecycle</a> in the <i>Amazon EC2 User Guide</i>.</p>
  *             </li>
  *          </ul>
  */
