+++ clients/client-ec2/commands/DescribePlacementGroupsCommand.ts
@@ -23,7 +23,7 @@ export type DescribePlacementGroupsCommandOutput = DescribePlacementGroupsResult
 /**
  * <p>Describes the specified placement groups or all of your placement groups. For more
  *             information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/placement-groups.html">Placement groups</a> in the
- *                 <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ *                 <i>Amazon EC2 User Guide</i>.</p>
  */
 export class DescribePlacementGroupsCommand extends $Command<
   DescribePlacementGroupsCommandInput,
