+++ clients/client-ec2/commands/DescribeReservedInstancesCommand.ts
@@ -23,7 +23,7 @@ export type DescribeReservedInstancesCommandOutput = DescribeReservedInstancesRe
 /**
  * <p>Describes one or more of the Reserved Instances that you purchased.</p>
  *          <p>For more information about Reserved Instances, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/concepts-on-demand-reserved-instances.html">Reserved
- * 				Instances</a> in the <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ * 				Instances</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class DescribeReservedInstancesCommand extends $Command<
   DescribeReservedInstancesCommandInput,
