+++ clients/client-ec2/commands/DescribeReservedInstancesListingsCommand.ts
@@ -26,7 +26,7 @@ export type DescribeReservedInstancesListingsCommandOutput = DescribeReservedIns
  * 		       <p>As a seller, you choose to list some or all of your Reserved Instances, and you specify the upfront price to receive for them. Your Reserved Instances are then listed in the Reserved Instance Marketplace and are available for purchase.</p>
  * 		       <p>As a buyer, you specify the configuration of the Reserved Instance to purchase, and the Marketplace matches what you're searching for with what's available. The Marketplace first sells the lowest priced Reserved Instances to you, and continues to sell available Reserved Instance listings to you until your demand is met. You are charged based on the total price of all of the listings that you purchase.</p>
  *          <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-market-general.html">Reserved Instance Marketplace</a>
- *         in the <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ *         in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class DescribeReservedInstancesListingsCommand extends $Command<
   DescribeReservedInstancesListingsCommandInput,
