+++ clients/client-ec2/commands/DescribeReservedInstancesModificationsCommand.ts
@@ -26,7 +26,7 @@ export type DescribeReservedInstancesModificationsCommandOutput = DescribeReserv
 
 /**
  * <p>Describes the modifications made to your Reserved Instances. If no parameter is specified, information about all your Reserved Instances modification requests is returned. If a modification ID is specified, only information about the specific modification is returned.</p>
- *    	     <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-modifying.html">Modifying Reserved Instances</a> in the Amazon Elastic Compute Cloud User Guide.</p>
+ *    	     <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-modifying.html">Modifying Reserved Instances</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class DescribeReservedInstancesModificationsCommand extends $Command<
   DescribeReservedInstancesModificationsCommandInput,
