+++ clients/client-ec2/commands/DescribeReservedInstancesOfferingsCommand.ts
@@ -28,7 +28,7 @@ export type DescribeReservedInstancesOfferingsCommandOutput = DescribeReservedIn
  * <p>Describes Reserved Instance offerings that are available for purchase. With Reserved Instances, you purchase the right to launch instances for a period of time. During that time period, you do not receive insufficient capacity errors, and you pay a lower usage rate than the rate charged for On-Demand instances for the actual time used.</p>
  *          <p>If you have listed your own Reserved Instances for sale in the Reserved Instance Marketplace, they will be excluded from these results. This is to ensure that you do not purchase your own Reserved Instances.</p>
  *          <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ri-market-general.html">Reserved Instance Marketplace</a>
- * 				in the <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ * 				in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class DescribeReservedInstancesOfferingsCommand extends $Command<
   DescribeReservedInstancesOfferingsCommandInput,
