+++ clients/client-ec2/commands/DisassociateRouteTableCommand.ts
@@ -1,5 +1,5 @@
 import { EC2ClientResolvedConfig, ServiceInputTypes, ServiceOutputTypes } from "../EC2Client";
-import { DisassociateRouteTableRequest } from "../models/models_4";
+import { DisassociateRouteTableRequest } from "../models/models_3";
 import {
   deserializeAws_ec2DisassociateRouteTableCommand,
   serializeAws_ec2DisassociateRouteTableCommand,
