+++ clients/client-ec2/commands/GetDefaultCreditSpecificationCommand.ts
@@ -23,8 +23,7 @@ export type GetDefaultCreditSpecificationCommandOutput = GetDefaultCreditSpecifi
 /**
  * <p>Describes the default credit option for CPU usage of a burstable performance instance family.</p>
  *         <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/burstable-performance-instances.html">Burstable
- *             performance instances</a> in the <i>Amazon Elastic Compute Cloud User
- *                 Guide</i>.</p>
+ *             performance instances</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class GetDefaultCreditSpecificationCommand extends $Command<
   GetDefaultCreditSpecificationCommandInput,
