+++ clients/client-ec2/commands/GetPasswordDataCommand.ts
@@ -22,7 +22,7 @@ export type GetPasswordDataCommandOutput = GetPasswordDataResult & __MetadataBea
  *         <p>The Windows password is generated at boot by the <code>EC2Config</code> service or
  *                 <code>EC2Launch</code> scripts (Windows Server 2016 and later). This usually only
  *             happens the first time an instance is launched. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/WindowsGuide/UsingConfig_WinAMI.html">EC2Config</a> and <a href="https://docs.aws.amazon.com/AWSEC2/latest/WindowsGuide/ec2launch.html">EC2Launch</a> in the
- *             Amazon Elastic Compute Cloud User Guide.</p>
+ *             <i>Amazon EC2 User Guide</i>.</p>
  *         <p>For the <code>EC2Config</code> service, the password is not generated for rebundled
  *             AMIs unless <code>Ec2SetPassword</code> is enabled before bundling.</p>
  *         <p>The password is encrypted using the key pair that you specified when you launched the
