+++ clients/client-ec2/commands/ModifyDefaultCreditSpecificationCommand.ts
@@ -33,8 +33,7 @@ export type ModifyDefaultCreditSpecificationCommandOutput = ModifyDefaultCreditS
  *             whether the update has occurred, you can call <code>GetDefaultCreditSpecification</code> and check
  *             <code>DefaultCreditSpecification</code> for updates.</p>
  *         <p>For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/burstable-performance-instances.html">Burstable
- *             performance instances</a> in the <i>Amazon Elastic Compute Cloud User
- *                 Guide</i>.</p>
+ *             performance instances</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class ModifyDefaultCreditSpecificationCommand extends $Command<
   ModifyDefaultCreditSpecificationCommandInput,
