+++ clients/client-ec2/commands/ModifyInstanceAttributeCommand.ts
@@ -30,8 +30,7 @@ export type ModifyInstanceAttributeCommandOutput = __MetadataBearer;
  *             associated with an ENI attached to an instance that has multiple ENIs, we recommend that
  *             you use the <a>ModifyNetworkInterfaceAttribute</a> action.</p>
  *         <p>To modify some attributes, the instance must be stopped. For more information, see
- *                 <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_ChangingAttributesWhileInstanceStopped.html">Modifying attributes of a stopped instance</a> in the <i>Amazon Elastic
- *                 Compute Cloud User Guide</i>.</p>
+ *                 <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Using_ChangingAttributesWhileInstanceStopped.html">Modifying attributes of a stopped instance</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class ModifyInstanceAttributeCommand extends $Command<
   ModifyInstanceAttributeCommandInput,
