+++ clients/client-ec2/commands/ModifyInstanceMetadataOptionsCommand.ts
@@ -26,7 +26,8 @@ export type ModifyInstanceMetadataOptionsCommandOutput = ModifyInstanceMetadataO
  *             started. When you modify the parameters on a running instance, the API responds with a
  *             state of “pending”. After the parameter modifications are successfully applied to the
  *             instance, the state of the modifications changes from “pending” to “applied” in
- *             subsequent describe-instances API calls. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-metadata.html">Instance metadata and user data</a>.</p>
+ *             subsequent describe-instances API calls. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-metadata.html">Instance metadata and user data</a>
+ *           in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class ModifyInstanceMetadataOptionsCommand extends $Command<
   ModifyInstanceMetadataOptionsCommandInput,
