+++ clients/client-ec2/commands/MonitorInstancesCommand.ts
@@ -23,8 +23,7 @@ export type MonitorInstancesCommandOutput = MonitorInstancesResult & __MetadataB
 /**
  * <p>Enables detailed monitoring for a running instance. Otherwise, basic monitoring is
  *             enabled. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-cloudwatch.html">Monitoring your instances and
- *                 volumes</a> in the <i>Amazon Elastic Compute Cloud User
- *             Guide</i>.</p>
+ *                 volumes</a> in the <i>Amazon EC2 User Guide</i>.</p>
  *         <p>To disable detailed monitoring, see .</p>
  */
 export class MonitorInstancesCommand extends $Command<
