+++ clients/client-ec2/commands/RebootInstancesCommand.ts
@@ -25,8 +25,7 @@ export type RebootInstancesCommandOutput = __MetadataBearer;
  *         <p>If an instance does not cleanly shut down within a few minutes, Amazon EC2 performs a
  *             hard reboot.</p>
  *         <p>For more information about troubleshooting, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-console.html">Getting console output and
- *                 rebooting instances</a> in the <i>Amazon Elastic Compute Cloud User
- *                 Guide</i>.</p>
+ *             rebooting instances</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class RebootInstancesCommand extends $Command<
   RebootInstancesCommandInput,
