+++ clients/client-ec2/commands/ResetInstanceAttributeCommand.ts
@@ -29,7 +29,7 @@ export type ResetInstanceAttributeCommandOutput = __MetadataBearer;
  *             checking is enabled. The default value is <code>true</code>, which means checking is
  *             enabled. This value must be <code>false</code> for a NAT instance to perform NAT. For
  *             more information, see <a href="https://docs.aws.amazon.com/AmazonVPC/latest/UserGuide/VPC_NAT_Instance.html">NAT Instances</a> in the
- *                 <i>Amazon Virtual Private Cloud User Guide</i>.</p>
+ *                 <i>Amazon VPC User Guide</i>.</p>
  */
 export class ResetInstanceAttributeCommand extends $Command<
   ResetInstanceAttributeCommandInput,
