+++ clients/client-ec2/commands/RevokeClientVpnIngressCommand.ts
@@ -1,5 +1,5 @@
 import { EC2ClientResolvedConfig, ServiceInputTypes, ServiceOutputTypes } from "../EC2Client";
-import { RevokeClientVpnIngressRequest, RevokeClientVpnIngressResult } from "../models/models_5";
+import { RevokeClientVpnIngressRequest, RevokeClientVpnIngressResult } from "../models/models_4";
 import {
   deserializeAws_ec2RevokeClientVpnIngressCommand,
   serializeAws_ec2RevokeClientVpnIngressCommand,
