+++ clients/client-ec2/commands/RunScheduledInstancesCommand.ts
@@ -26,7 +26,7 @@ export type RunScheduledInstancesCommandOutput = RunScheduledInstancesResult & _
  *          <p>You must launch a Scheduled Instance during its scheduled time period. You can't stop or reboot a Scheduled Instance,
  *          but you can terminate it as needed. If you terminate a Scheduled Instance before the current scheduled time period ends,
  *          you can launch it again after a few minutes. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-scheduled-instances.html">Scheduled Instances</a>
- *          in the <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ *          in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class RunScheduledInstancesCommand extends $Command<
   RunScheduledInstancesCommandInput,
