+++ clients/client-ec2/commands/StopInstancesCommand.ts
@@ -22,7 +22,7 @@ export type StopInstancesCommandOutput = StopInstancesResult & __MetadataBearer;
  *         <p>You can use the Stop action to hibernate an instance if the instance is <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Hibernate.html#enabling-hibernation">enabled for
  *                 hibernation</a> and it meets the <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Hibernate.html#hibernating-prerequisites">hibernation
  *                 prerequisites</a>. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/Hibernate.html">Hibernate your instance</a> in the
- *                 <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ *                 <i>Amazon EC2 User Guide</i>.</p>
  *         <p>We don't charge usage for a stopped instance, or data transfer fees; however, your
  *             root partition Amazon EBS volume remains and continues to persist your data, and you are
  *             charged for Amazon EBS volume usage. Every time you start your Windows instance, Amazon
@@ -33,8 +33,7 @@ export type StopInstancesCommandOutput = StopInstancesResult & __MetadataBearer;
  *             usage, and thereafter charges per second for instance usage.</p>
  *         <p>You can't stop or hibernate instance store-backed instances. You can't use the Stop
  *             action to hibernate Spot Instances, but you can specify that Amazon EC2 should hibernate
- *             Spot Instances when they are interrupted. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/spot-interruptions.html#hibernate-spot-instances">Hibernating interrupted Spot Instances</a> in the <i>Amazon Elastic
- *                 Compute Cloud User Guide</i>.</p>
+ *             Spot Instances when they are interrupted. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/spot-interruptions.html#hibernate-spot-instances">Hibernating interrupted Spot Instances</a> in the <i>Amazon EC2 User Guide</i>.</p>
  *         <p>When you stop or hibernate an instance, we shut it down. You can restart your instance
  *             at any time. Before stopping or hibernating an instance, make sure it is in a state from
  *             which it can be restarted. Stopping an instance does not preserve data stored in RAM,
@@ -46,12 +45,11 @@ export type StopInstancesCommandOutput = StopInstancesResult & __MetadataBearer;
  *             any other devices attached during the instance launch are automatically deleted. For
  *             more information about the differences between rebooting, stopping, hibernating, and
  *             terminating instances, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-lifecycle.html">Instance lifecycle</a>
- *             in the <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
+ *             in the <i>Amazon EC2 User Guide</i>.</p>
  *         <p>When you stop an instance, we attempt to shut it down forcibly after a short while. If
  *             your instance appears stuck in the stopping state after a period of time, there may be
  *             an issue with the underlying host computer. For more information, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/TroubleshootingInstancesStopping.html">Troubleshooting
- *                 stopping your instance</a> in the <i>Amazon Elastic Compute Cloud User
- *                 Guide</i>.</p>
+ *                 stopping your instance</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class StopInstancesCommand extends $Command<
   StopInstancesCommandInput,
