+++ clients/client-ec2/commands/TerminateInstancesCommand.ts
@@ -36,9 +36,8 @@ export type TerminateInstancesCommandOutput = TerminateInstancesResult & __Metad
  *             EBS volumes with the <code>DeleteOnTermination</code> block device mapping parameter set
  *             to <code>true</code> are automatically deleted. For more information about the
  *             differences between stopping and terminating instances, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-instance-lifecycle.html">Instance lifecycle</a>
- *             in the <i>Amazon Elastic Compute Cloud User Guide</i>.</p>
- *         <p>For more information about troubleshooting, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/TroubleshootingInstancesShuttingDown.html">Troubleshooting terminating your instance</a> in the <i>Amazon Elastic
- *                 Compute Cloud User Guide</i>.</p>
+ *             in the <i>Amazon EC2 User Guide</i>.</p>
+ *         <p>For more information about troubleshooting, see <a href="https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/TroubleshootingInstancesShuttingDown.html">Troubleshooting terminating your instance</a> in the <i>Amazon EC2 User Guide</i>.</p>
  */
 export class TerminateInstancesCommand extends $Command<
   TerminateInstancesCommandInput,
