+++ clients/client-ec2/index.ts
@@ -163,7 +163,6 @@ export * from "./commands/DeregisterTransitGatewayMulticastGroupMembersCommand";
 export * from "./commands/DeregisterTransitGatewayMulticastGroupSourcesCommand";
 export * from "./commands/DescribeAccountAttributesCommand";
 export * from "./commands/DescribeAddressesCommand";
-export * from "./commands/DescribeAddressesAttributeCommand";
 export * from "./commands/DescribeAggregateIdFormatCommand";
 export * from "./commands/DescribeAvailabilityZonesCommand";
 export * from "./commands/DescribeBundleTasksCommand";
@@ -451,7 +450,6 @@ export * from "./commands/ImportInstanceCommand";
 export * from "./commands/ImportKeyPairCommand";
 export * from "./commands/ImportSnapshotCommand";
 export * from "./commands/ImportVolumeCommand";
-export * from "./commands/ModifyAddressAttributeCommand";
 export * from "./commands/ModifyAvailabilityZoneGroupCommand";
 export * from "./commands/ModifyCapacityReservationCommand";
 export * from "./commands/ModifyClientVpnEndpointCommand";
@@ -522,7 +520,6 @@ export * from "./commands/ReplaceTransitGatewayRouteCommand";
 export * from "./commands/ReportInstanceStatusCommand";
 export * from "./commands/RequestSpotFleetCommand";
 export * from "./commands/RequestSpotInstancesCommand";
-export * from "./commands/ResetAddressAttributeCommand";
 export * from "./commands/ResetEbsDefaultKmsKeyIdCommand";
 export * from "./commands/ResetFpgaImageAttributeCommand";
 export * from "./commands/ResetImageAttributeCommand";
